Moves goods rules to goods app
Fix the error related to app registration and move the rules
