+++ core/goods/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CoreGoodsConfig(AppConfig):
+    name = "core.goods"
+    label = "core_goods"
