+++ exporter/goods/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class GoodsConfig(AppConfig):
+    name = "exporter.goods"
