+++ exporter/goods/services.py
@@ -14,6 +14,7 @@ def get_goods(
     data = client.get(request, "/goods/" + convert_parameters_to_query_params(locals()))
     return data.json()
 
+
 def get_archived_goods(
     request, page: int = 1, name=None, part_number=None, control_list_entry=None, for_application=None
 ):
